Drill

+++ services/api.js
@@ -56,34 +56,34 @@ export const uploadProfilePhotoAPI = (formData) =>
             'User-Agent': 'PostmanRuntime/7.41.1',
         },
     });
-export const toggleFavourite = (videoId, isFavourite) => {
-    return API.put(`/video/${videoId}`, { isFavourite });
+export const toggleFavourite = (videoId, params) => {
+    return API.put(`/video/${videoId}`, params);
 };
 export const uploadVideo = (formData) =>
-  API.post('/videos/upload', formData, {
-    headers: { 'Content-Type': 'multipart/form-data' }
-  });
+    API.post('/videos/upload', formData, {
+        headers: { 'Content-Type': 'multipart/form-data' }
+    });
 // Use axios for annotations
 export const addAnnotation = (videoId, annotationData, token) =>
-  API.post(`/video-ann/${videoId}/annotations`, annotationData, {
-    headers: { Authorization: `Bearer ${token}` }
-  });
+    API.post(`/video-ann/${videoId}/annotations`, annotationData, {
+        headers: { Authorization: `Bearer ${token}` }
+    });
 
 export const getAnnotations = (videoId, token) =>
-  API.get(`/video-ann/${videoId}/annotations`, {
-    headers: { Authorization: `Bearer ${token}` }
-  });
+    API.get(`/video-ann/${videoId}/annotations`, {
+        headers: { Authorization: `Bearer ${token}` }
+    });
 
 export const updateAnnotation = (videoId, annotationId, annotationData, token) =>
-  API.put(`/video-ann/${videoId}/annotations/${annotationId}`, annotationData, {
-    headers: { Authorization: `Bearer ${token}` }
-  });
+    API.put(`/video-ann/${videoId}/annotations/${annotationId}`, annotationData, {
+        headers: { Authorization: `Bearer ${token}` }
+    });
 
 export const deleteAnnotation = (videoId, annotationId, token) =>
-  API.delete(`/video-ann/${videoId}/annotations/${annotationId}`, {
-    headers: { Authorization: `Bearer ${token}` }
-  });
+    API.delete(`/video-ann/${videoId}/annotations/${annotationId}`, {
+        headers: { Authorization: `Bearer ${token}` }
+    });
 
-export const deleteVideos = (videoIds) =>
-    API.delete('/video/delete', { data: { ids: videoIds } });
+export const deleteVideos = (videoIds, userId) =>
+    API.delete('/video/delete', { data: { ids: videoIds, userId } });
 export default API;
